fixing build issues after rebase
