+++ p2p/helpers.go
@@ -1,10 +1,10 @@
 package p2p
 
 import (
-	"github.com/spacemeshos/go-spacemesh/p2p/net"
 	"github.com/spacemeshos/go-spacemesh/p2p/nodeconfig"
 	"github.com/stretchr/testify/assert"
 	"testing"
+	"github.com/spacemeshos/go-spacemesh/p2p/node"
 )
 
 func defaultConfig() nodeconfig.Config {
@@ -12,7 +12,7 @@ func defaultConfig() nodeconfig.Config {
 }
 
 func p2pTestInstance(t testing.TB, config nodeconfig.Config) Swarm {
-	port, err := net.GetUnboundedPort()
+	port, err := node.GetUnboundedPort()
 	assert.NoError(t, err, "Error getting a port", err)
 	config.TCPPort = port
 	p, err := New(config, false)
